Simplify reaper to cgroup-only (assume systemd is always available)
The deployment host always has a systemd user session with cgroup
delegation, and the process-group fallback was strictly weaker (a
double-forked grandchild escapes a process-group kill, so only the cgroup
mode reaps the motivating Postgres case). Drop the fallback mode, the
capability probe, the memoization, and the ReaperMode type:

- wrapBackgroundCommand(command) now emits only the systemd user-scope
  wrapper; computeBackgroundBashRewrite loses its mode param.
- backgroundReaperHook calls the rewrite directly (no detectReaperMode,
  no execFile probe); it now builds the wrapper string without spawning,
  so its tests move to a fast unit test (background-reaper.test.ts).
- Integration test keeps the real double-fork cgroup-reaping and
  stdout/exit-code preservation checks (skipped where no user scope, e.g.
  CI); the process-group test is removed with the mode.

Net ~150 fewer lines. Also considered running the whole session in one
cgroup instead of the hook, but the SDK's per-session subprocess launch
can't be interposed cleanly (its default executable path isn't stably
obtainable) and a session cgroup only reaps at teardown, not per killed
background task — a weaker fix. Docs updated.
