+++ doc/DESIGN.md
@@ -548,12 +548,11 @@ When the SDK terminates a **still-running** `run_in_background` **Bash** task (i
 
 The app fixes this with a **`PreToolUse` hook** (`backgroundReaperHook` in [`src/server/services/background-reaper.ts`](../src/server/services/background-reaper.ts), wired in `buildSdkOptions`). For a backgrounded Bash call it rewrites the tool's `command` (via the hook's `updatedInput`) to run under a supervisor that reaps the **whole process tree** on SIGTERM/SIGINT/SIGHUP. (That the SDK honors a `PreToolUse` `updatedInput` for Bash — the rewritten command runs, not the original — was verified with a live query in `scripts/spike-pretooluse-rewrite.ts`.) Only `run_in_background` Bash is wrapped; every other tool call passes through untouched, and the hook **fails open** (any error runs the command unwrapped). The command-building is the pure, unit-tested `wrapBackgroundCommand` in [`src/lib/background-command.ts`](../src/lib/background-command.ts); the real reaping is covered by `background-reaper.integration.test.ts` (spawns a double-forking daemon and asserts it dies).
 
-Two teardown mechanisms, chosen by a one-time capability probe (`detectReaperMode`, memoized):
+**Teardown mechanism.** The command runs inside a transient **systemd user scope** (`systemd-run --user --scope` — its own cgroup) and the supervisor's trap does `systemctl --user stop <scope>`. A cgroup stop kills every process in the tree regardless of double-forking or `setsid`, so it reaps the daemonized grandchild. This is the unprivileged-cgroup approach; it assumes a systemd **user** session with cgroup delegation is available — true on the deployment host (if `systemd-run` were ever missing, the wrapped command fails loudly rather than running unreaped). The scope is given a short `TimeoutStopSec` so a process that ignores SIGTERM can't wedge the stop near systemd's 90s default before the cgroup SIGKILL. The original command is passed to the scope **base64-encoded as a positional arg** and decoded (`base64 -d <<<"$1" | bash`) _inside_ the scope: systemd applies its own `ExecStart` expansion to the command line (`$$` → `$`, `$VAR` → env), so any `$` in the script must be kept out of systemd's view — base64's alphabet has no `$`, and `$1` is digit-led so systemd doesn't treat it as a variable.
 
-- **`cgroup` (preferred).** The command runs inside a transient **systemd user scope** (`systemd-run --user --scope`) — its own cgroup — and the supervisor's trap does `systemctl --user stop <scope>`. A cgroup stop kills every process in the tree regardless of double-forking or `setsid`, so it reaps the daemonized grandchild. This is the unprivileged-cgroup approach; it requires a systemd **user** session with cgroup delegation, which the probe (`systemd-run --user --scope -- true`) detects. The original command is passed to the scope **base64-encoded as a positional arg** and decoded (`base64 -d <<<"$1" | bash`) _inside_ the scope: systemd applies its own `ExecStart` expansion to the command line (`$$` → `$`, `$VAR` → env), so any `$` in the script must be kept out of systemd's view — base64's alphabet has no `$`, and `$1` is digit-led so systemd doesn't treat it as a variable.
-- **`process-group` (fallback).** When no user scope is available, the command runs as a job in its own process group (`set -m`, so the job's pid == its pgid) and the trap signals the whole group: SIGTERM (a grace period lets the command's own shell `trap` handlers run — e.g. the common `pnpm services` case where the trap does `pg_ctl stop`), then SIGKILL. A grandchild that double-forked out of the group still escapes this mode — best-effort only.
+(An earlier version also carried a `process-group` fallback — `set -m` + signal the group — for hosts without a user scope. It was dropped because the deployment host always has systemd, and it was strictly weaker: a double-forked grandchild escapes a process-group kill, so only the cgroup mode actually reaps the motivating Postgres case.)
 
-The wrapper preserves the command's stdout/stderr and exit code (verified empirically for both modes), so the SDK's `BashOutput` view and completion status are unchanged. _Known limitation_: an untrappable SIGKILL to the supervisor can't run the teardown, so a `cgroup` scope would linger until its own processes exit; the observed termination is the trappable SIGTERM.
+The wrapper preserves the command's stdout/stderr and exit code (verified empirically), so the SDK's `BashOutput` view and completion status are unchanged. _Known limitation_: an untrappable SIGKILL to the supervisor can't run the teardown, so the scope would linger until its own processes exit; the observed termination is the trappable SIGTERM.
 
 ### Cost & Context Usage Estimation
